[arcane] Nettoyage et déplace 'MeshComponentPartData' dans 'arcane/core/materials'.
